fix(preprocess): dtype guard in mcap_neutralize_panel + strengthen independence test
Quality-review follow-up on dd7b3c0:
- df.copy() → df.astype(float).copy() to prevent silent int→float downcast
  if an integer-typed factor panel ever flows in.
- test_days_are_independent now asserts the mutated day IS different,
  catching trivial-pass failure modes where the function ignores log_mcap.

+++ src/stockpool/ml/preprocess.py
@@ -146,7 +146,7 @@ def mcap_neutralize_panel(
     if df.empty:
         return df.copy()
     log_mcap_aligned = log_mcap.reindex(index=df.index, columns=df.columns)
-    out = df.copy()
+    out = df.astype(float).copy()
     fallback_days = 0
     for date in df.index:
         y = df.loc[date]
